Get rid of example_run.py and example_plot.py
